+++ src/features/Leaderboards/index.js
@@ -24,11 +24,13 @@ const Leaderboards = () => (
         name: 'PvP (NA)',
         content: <PvpLeaderboard region="na" />,
         description: config.descriptions.pvpLeaderboard,
+        flair: 'new',
       }, {
         to: '/leaderboards/pvp/eu',
         name: 'PvP (EU)',
         content: <PvpLeaderboard region="eu" />,
         description: config.descriptions.pvpLeaderboard,
+        flair: 'new',
       }]}
     />